Narrow Replica from usize to u16
Two-byte replica id lets libnet-rs receivers wrap the wire payload with
a sender field at trivial overhead, and fixes two pre-existing typos
(apollo's round() returning Replica instead of Round, and artemis's
myid() returning usize instead of Replica) that were silently OK while
Replica aliased usize.

u16 comfortably covers BFT deployments (up to 65k nodes). Broadcast
sentinel sites (num_nodes-as-recipient) cast explicitly.

+++ types/src/protocol.rs
@@ -1,2 +1,2 @@
-pub type Replica = usize;
+pub type Replica = u16;
 pub type Height = usize;
